+++ CHANGELOG.md
@@ -5,13 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.0] - 2026-02-24
 
 ### Added
 
 - Geocoding of weather stations used in Met Éireann datasets
 - Ability to download and merge ungeneralised geometries from Tailte Éireann/OSNI
 - NOTICES.md containing third-party copyright notices and license information
+
+### Fixed
+
+- Prevent pivot methods from dropping statistic columns that become all-NaN after filtering
+- Added alternative filecode for ungeneralised 2017 constituency geometries
 
 ## [0.1.0] - 2026-02-03